add ansible

+++ build/build.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
+dir=${PWD##*/}
+build_dir="build"
+if [ "$dir" != "$build_dir" ]; then
+  exit 1
+fi
 for i in $(ls | grep -v build.sh); do
 	rm -Rf ./$i
 done
